Moved the async update strategies into their own file to keep dvi.py clean.
